Update header-entete.php
Fixed the PHP header to use the svg images and fall back to the png images. It seems to work, I spent half an hour trying to get the svg image to show up when I realized the server has to have the svg mimetype added to it for the svg image to work.