WET-BOEW-PHP Issue 54 - Warning image missing for Archived Content message
This patch resolves an issue where warning image is missing for Archived Content message if /wet-boew/dist folder is not in root of website or has a different name.

Also adds the missing $_SITE['archived_'.$_LANG_] line to the non-English and non-French versions so that it doesn't produce an error.

Someone else will need to provide appropriate translations.

Signed-off-by: Michael Milette <michael.milette@tngconsulting.ca>