Added custom URL support to language switcher.
If URL's of some applications don't end in '-en' or '-fr', for example, but uses '?lang=en' to switch language. In this case, the script didn't work. This fix will enable applications to provide their own correct full URL and not depend on the built-in mechanism.

Example of URL that fails without this fix: http://www.example.com/page/?lang=en

This fix re-enables functionality which had been previously added to WET-BOEW-PHP v3.1.