Update head-css.php
Removed a test to see if the page is a splash page (usually index.php), The if statement originally prevented the inclusion of the /wet-boew.min.css file, but as it turns out the file is required to render buttons properly. I think at some point they must have factored the buttons out of the /theme-sp-pe.css file into the /wet-boew.min.css file.